make text drawing algorithm a util function
